docs: add instance for movie-web.x / IPFS / IPNS
IPFS is a decentralized way to store, distribute and access static content. Perfect for the frontend.

Unstoppable Domains are domains stored on the polygon blockchain which are exclusively controlled by the person holding the private key for the domain.

Both systems are censorship resistant.

Gateway usage should be generally discouraged over native use as gateways may block access to certain content / names.